requirements.txt
accelerate==1.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.0
annotated-types==0.7.0
astroid==3.3.8
asttokens==3.0.0
attrs==24.3.0
banal==1.0.6
bitsandbytes==0.45.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
dataset==1.6.2
datasets==3.2.0
debugpy==1.8.12
decorator==5.1.1
dill==0.3.8
docker-pycreds==0.4.0
executing==2.2.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-genai==1.2.0
greenlet==3.1.1
huggingface-hub==0.27.0
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.5
isort==6.0.0
jedi==0.19.2
Jinja2==3.1.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pandas==2.2.3
parso==0.8.4
peft==0.14.0
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.2
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.19.0
pylint==3.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rsa==4.9
safetensors==0.5.0
sentry-sdk==2.19.2
setproctitle==1.3.4
setuptools==75.6.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==1.4.54
stack-data==0.6.3
sympy==1.13.1
tokenizers==0.21.0
tomlkit==0.13.2
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
triton==3.1.0
trl==0.13.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wandb==0.19.1
wcwidth==0.2.13
websockets==14.2
widgetsnbextension==4.0.13
xxhash==3.5.0
yarl==1.18.3
